+++ timelock-agent/src/main/java/com/palantir/timelock/config/PaxosInstallConfiguration.java
@@ -22,7 +22,6 @@ import org.immutables.value.Value;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.google.common.base.Preconditions;
 
 @JsonDeserialize(as = ImmutablePaxosInstallConfiguration.class)
 @JsonSerialize(as = ImmutablePaxosInstallConfiguration.class)
@@ -62,8 +61,5 @@ public interface PaxosInstallConfiguration {
                             + "nodes, you have likely already made a mistake by this point. This is a non-trivial "
                             + "operation, please be careful and make sure that you have appropriate approvals.");
         }
-
-        Preconditions.checkArgument(dataDirectory().mkdirs() || dataDirectory().isDirectory(),
-                "Could not create paxos data directory %s", dataDirectory());
     }
 }